LayoutTests/http/tests/performance/paint-timing/performance-paint-timing-fcp-after-visually-non-empty-for-style-expected.txt

PASS paintEntry.startTime >= styleEntry.startTime + styleEntry.duration is true
TEXT
